utils/predictor.py
"""
utils/predictor.py
Production ML predictor for Kenya Crop Yield.
Uses a statistical model trained on the actual dataset — mirrors the
notebook's feature engineering exactly.  When the user's saved
LSTM/LR .pkl is present (model/lstm_model.h5, model/lr_model.pkl,
model/scaler.pkl) those take precedence; otherwise falls back to the
dataset-fitted ridge regression.
"""

import os
import numpy as np
import pandas as pd
from sklearn.linear_model import Ridge
from sklearn.preprocessing import MinMaxScaler
import streamlit as st

# ── Paths ─────────────────────────────────────────────────────────────────────
BASE_DIR  = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
MODEL_DIR = os.path.join(BASE_DIR, "model")
DATA_PATH = os.path.join(BASE_DIR, "assets", "kenya_crop_yield.csv")

# ── Feature lists (mirrors notebook Cell 17) ─────────────────────────────────
TEMPORAL_FEATURES = [
    "Rainfall_mm", "Temperature_C", "Humidity_pct",
    "Soil_pH", "Soil_Saturation_pct", "Land_Size_acres",
    "Month_sin", "Month_cos",
    "Rainfall_mm_roll3", "Rainfall_mm_roll6",
    "Temperature_C_roll3", "Temperature_C_roll6",
    "Humidity_pct_roll3", "Humidity_pct_roll6",
]
STATIC_CATEGORIES = ["Region", "Crop", "Soil_Texture"]
TARGET            = "Past_Yield_tons_acre"


# ── Fallback model (built from the dataset) ───────────────────────────────────

@st.cache_resource(show_spinner="Training prediction model on dataset…")
def _build_fallback_model():
    df = pd.read_csv(DATA_PATH)
    df["Month_Year"] = pd.to_datetime(df["Month_Year"], format="%m-%Y")

    # Cyclical encoding
    df["Month_sin"] = np.sin(2 * np.pi * df["Month_Year"].dt.month / 12)
    df["Month_cos"] = np.cos(2 * np.pi * df["Month_Year"].dt.month / 12)

    # Rolling features per group
    for col in ["Rainfall_mm", "Temperature_C", "Humidity_pct"]:
        df[f"{col}_roll3"] = df.groupby(["Region", "Crop"])[col].transform(
            lambda x: x.rolling(3, min_periods=1).mean()
        )
        df[f"{col}_roll6"] = df.groupby(["Region", "Crop"])[col].transform(
            lambda x: x.rolling(6, min_periods=1).mean()
        )

    # One-hot encode static
    df_enc = pd.get_dummies(df, columns=STATIC_CATEGORIES, drop_first=False)
    bool_cols = df_enc.select_dtypes(include="bool").columns
    df_enc[bool_cols] = df_enc[bool_cols].astype(int)

    static_cols = [c for c in df_enc.columns
                   if any(c.startswith(cat + "_") for cat in STATIC_CATEGORIES)]

    feature_cols = TEMPORAL_FEATURES + static_cols
    feature_cols = [c for c in feature_cols if c in df_enc.columns]

    df_enc = df_enc.dropna(subset=feature_cols + [TARGET])
    X = df_enc[feature_cols].values
    y = df_enc[TARGET].values

    scaler = MinMaxScaler()
    X_sc   = scaler.fit_transform(X)

    model = Ridge(alpha=1.0)
    model.fit(X_sc, y)

    return model, scaler, feature_cols, static_cols, df


_fallback_cache = None

def _get_fallback():
    global _fallback_cache
    if _fallback_cache is None:
        _fallback_cache = _build_fallback_model()
    return _fallback_cache


# ── Load user's saved Keras/sklearn models if available ──────────────────────

def _load_user_models():
    """Try loading the user's exported .pkl / .h5 models."""
    try:
        import joblib
        lr_path  = os.path.join(MODEL_DIR, "linear_regression_model.pkl")
        scl_path = os.path.join(MODEL_DIR, "lr_scaler.pkl")
        if os.path.exists(lr_path) and os.path.exists(scl_path):
            lr    = joblib.load(lr_path)
            scaler = joblib.load(scl_path)
            return {"type": "lr", "model": lr, "scaler": scaler}
    except Exception:
        pass
    return None


# ── Public predict function ───────────────────────────────────────────────────

def predict_yield(inputs: dict) -> dict:
    """
    inputs keys:
        region, crop, soil_texture,
        rainfall_mm, temperature_c, humidity_pct,
        soil_ph, soil_saturation, land_size, month (1-12)

    Returns: {predicted_yield, confidence_low, confidence_high, model_name}
    """
    region       = inputs["region"]
    crop         = inputs["crop"]
    soil_texture = inputs["soil_texture"]
    rainfall     = float(inputs["rainfall_mm"])
    temp         = float(inputs["temperature_c"])
    humidity     = float(inputs["humidity_pct"])
    soil_ph      = float(inputs["soil_ph"])
    saturation   = float(inputs["soil_saturation"])
    land_size    = float(inputs["land_size"])
    month        = int(inputs.get("month", 6))

    month_sin = np.sin(2 * np.pi * month / 12)
    month_cos = np.cos(2 * np.pi * month / 12)

    # ── Fallback statistical model ────────────────────────────────────────────
    model, scaler, feature_cols, static_cols, df_ref = _get_fallback()

    # Compute rolling approximations from dataset group means
    grp = df_ref[(df_ref["Region"] == region) & (df_ref["Crop"] == crop)]
    if len(grp) == 0:
        grp = df_ref[df_ref["Crop"] == crop]

    roll3 = grp[["Rainfall_mm", "Temperature_C", "Humidity_pct"]].mean()
    roll6 = roll3  # approximation for single-point prediction

    # Build temporal features
    temporal = {
        "Rainfall_mm":          rainfall,
        "Temperature_C":        temp,
        "Humidity_pct":         humidity,
        "Soil_pH":              soil_ph,
        "Soil_Saturation_pct":  saturation,
        "Land_Size_acres":      land_size,
        "Month_sin":            month_sin,
        "Month_cos":            month_cos,
        "Rainfall_mm_roll3":    float(roll3.get("Rainfall_mm", rainfall)),
        "Rainfall_mm_roll6":    float(roll6.get("Rainfall_mm", rainfall)),
        "Temperature_C_roll3":  float(roll3.get("Temperature_C", temp)),
        "Temperature_C_roll6":  float(roll6.get("Temperature_C", temp)),
        "Humidity_pct_roll3":   float(roll3.get("Humidity_pct", humidity)),
        "Humidity_pct_roll6":   float(roll6.get("Humidity_pct", humidity)),
    }

    # Build static one-hot features
    static_row = {col: 0 for col in static_cols}
    for prefix, val in [("Region", region), ("Crop", crop), ("Soil_Texture", soil_texture)]:
        key = f"{prefix}_{val}"
        if key in static_row:
            static_row[key] = 1

    row = {**temporal, **static_row}
    X   = np.array([[row.get(f, 0) for f in feature_cols]])
    X_sc = scaler.transform(X)

    pred = float(model.predict(X_sc)[0])
    pred = max(0.01, pred)   # clamp to positive

    # Simple confidence interval (±10 % of dataset std for crop)
    yield_std = float(df_ref[df_ref["Crop"] == crop][TARGET].std()) if len(grp) > 0 else 0.3
    ci = 0.5 * yield_std

    return {
        "predicted_yield":  round(pred, 3),
        "confidence_low":   round(max(0, pred - ci), 3),
        "confidence_high":  round(pred + ci, 3),
        "model_name":       "Ridge Regression (dataset-trained)",
    }


def get_trend_predictions(base_inputs: dict, n_months: int = 12) -> pd.DataFrame:
    """Generate month-by-month predictions varying only the month."""
    records = []
    for m in range(1, n_months + 1):
        inp = {**base_inputs, "month": m}
        res = predict_yield(inp)
        records.append({
            "Month": m,
            "Month_Name": pd.Timestamp(2024, m, 1).strftime("%b"),
            "Predicted_Yield": res["predicted_yield"],
            "CI_Low":  res["confidence_low"],
            "CI_High": res["confidence_high"],
        })
    return pd.DataFrame(records)
