requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
scikit-learn==1.4.2
supabase==2.4.6
python-dotenv==1.0.1
bcrypt==4.1.3
reportlab==4.2.0
fpdf2==2.7.9
Pillow==10.3.0
streamlit-option-menu==0.3.12
